+++ lingui.config.ts
@@ -20,6 +20,14 @@ const config: LinguiConfig = {
       include: ['packages/attention/index.js'],
       path: 'packages/attention/locales/{locale}/messages',
     },
+    {
+      include: ['packages/toast/toast.js'],
+      path: 'packages/toast/locales/{locale}/messages',
+    },
+    {
+      include: ['packages/pill/index.js'],
+      path: 'packages/pill/locales/{locale}/messages',
+    },
   ],
   compileNamespace: 'es',
   sourceLocale: 'en',