+++ packages/toast/toast.js
@@ -4,10 +4,15 @@ import { classMap } from 'lit/directives/class-map.js';
 import { when } from 'lit/directives/when.js';
 import { toast as ccToast } from '@warp-ds/css/component-classes';
 import { expand, collapse } from 'element-collapse';
-import '@warp-ds/icons/elements/alert-warning-16'
-import '@warp-ds/icons/elements/alert-error-16'
-import '@warp-ds/icons/elements/alert-success-16'
+import '@warp-ds/icons/elements/warning-16'
+import '@warp-ds/icons/elements/error-16'
+import '@warp-ds/icons/elements/success-16'
 import '@warp-ds/icons/elements/close-16'
+import { i18n } from '@lingui/core'
+import { messages as enMessages } from './locales/en/messages.mjs'
+import { messages as nbMessages } from './locales/nb/messages.mjs'
+import { messages as fiMessages } from './locales/fi/messages.mjs'
+import { activateI18n } from '../i18n'
 
 const classes = (definition) => {
   const defn = {};
@@ -43,6 +48,8 @@ export class WarpToast extends WarpElement {
 
   constructor() {
     super();
+    activateI18n(enMessages, nbMessages, fiMessages);
+
     this.id = Date.now().toString(36) + Math.random().toString(36).slice(2, 5);
     this.type = 'success';
     this.text = '';
@@ -96,15 +103,30 @@ export class WarpToast extends WarpElement {
   }
 
   get _typeLabel() {
-    if (this._warning) return 'Varsel';
-    if (this._error) return 'Feil';
-    return 'Vellykket'
+    if (this._warning) return i18n._({
+      id: 'toast.aria.warning',
+      message: 'Warning',
+      comment:
+        'Default screenreader message for warning in toast component',
+    });
+    if (this._error) return i18n._({
+      id: 'toast.aria.error',
+      message: 'Error',
+      comment:
+        'Default screenreader message for error in toast component',
+    });
+    return i18n._({
+      id: 'toast.aria.successful',
+      message: 'Successful',
+      comment:
+        'Default screenreader message for successful in toast component',
+    });
   }
 
   get _iconMarkup() {
-    if (this._warning) return html`<w-icon-alert-warning-16></w-icon-alert-warning-16>`;
-    if (this._error) return html`<w-icon-alert-error-16></w-icon-alert-error-16>`;
-    else return html`<w-icon-alert-success-16></w-icon-alert-success-16>`;
+    if (this._warning) return html`<w-icon-warning-16></w-icon-warning-16>`;
+    if (this._error) return html`<w-icon-error-16></w-icon-error-16>`;
+    else return html`<w-icon-success-16></w-icon-success-16>`;
   }
 
   async collapse() {